Fix User Update screen + User operations functions

+++ lib/Screens/Login_register/register_screen.dart
@@ -247,7 +247,7 @@ class _RegisterScreenState extends State<RegisterScreen> {
                         if (value == null || value.isEmpty) {
                           return 'Current Weight field is required';
                         }
-                        else if ((!RegExp(r'^[0-9]+$').hasMatch(value)) || (int.parse(value) < 5) || (!RegExp(r'^[0]+[0-9]*$').hasMatch(value))){
+                        else if ((!RegExp(r'^[0-9]+$').hasMatch(value)) || (double.parse(value) < 5) || (!RegExp(r'^[0]+[0-9]*$').hasMatch(value))){
                           return 'Please enter a valid weight';
                         }
                         return null;
@@ -275,7 +275,7 @@ class _RegisterScreenState extends State<RegisterScreen> {
                         if (value == null || value.isEmpty) {
                           return 'Goal Weight field is required';
                         }
-                        else if ((!RegExp(r'^[0-9]+$').hasMatch(value)) || (int.parse(value) < 5) || (!RegExp(r'^[0]+[0-9]*$').hasMatch(value))){
+                        else if ((!RegExp(r'^[0-9]+$').hasMatch(value)) || (double.parse(value) < 5) || (!RegExp(r'^[0]+[0-9]*$').hasMatch(value))){
                           return 'Please enter a valid weight';
                         }
                         return null;
@@ -303,7 +303,7 @@ class _RegisterScreenState extends State<RegisterScreen> {
                         if (value == null || value.isEmpty) {
                           return 'Height field is required';
                         }
-                        else if ((!RegExp(r'^[0-9]+$').hasMatch(value)) || (int.parse(value) < 40) || (int.parse(value) > 300) || (!RegExp(r'^[0]+[0-9]*$').hasMatch(value))){
+                        else if ((!RegExp(r'^[0-9]+$').hasMatch(value)) || double.parse(value) < 40 || double.parse(value) > 300 || (!RegExp(r'^[0]+[0-9]*$').hasMatch(value))){
                           return 'Please enter a valid height';
                         }
                         return null;
